requirements.txt
pandas==1.3.5
tqdm==4.63.0
scikit-learn==1.0.2
numpy==1.21.5
scikit-image==0.19.2
Pillow==9.0.1
matplotlib==3.5.1
PyYAML==6.0
scipy==1.7.3
openpyxl==3.0.9
torch==1.10.2
torchvision==0.11.3